+++ src/main/java/com/octl3/api/dto/CertificateDto.java
@@ -14,8 +14,8 @@ import java.time.LocalDate;
 @AllArgsConstructor
 @JsonInclude(JsonInclude.Include.NON_NULL)
 public class CertificateDto {
-    private Integer id;
-    private Integer employeeId;
+    private Long id;
+    private Long employeeId;
     private String name;
     private String field;
     private String content;
